requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
opencv-python-headless==4.10.0.84
httpx==0.27.0
numpy==1.26.4
Pillow==10.4.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pydantic[email]==2.9.2
